refactor(workspace): split combat, items, and games modules
game/src/tributes/combat/mod.rs -> mod.rs (orchestrator) + resolve.rs (combat resolution) + stats.rs (update_stats) + tests.rs (all tests)

game/src/items/mod.rs -> mod.rs (Item, ItemType) + generation.rs (random item generation) + tests.rs (item tests)

api/src/games.rs -> api/src/games/mod.rs (handler routers) + handlers.rs (handler implementations)
