+++ game/src/tributes/combat/mod.rs
@@ -7,65 +7,31 @@
 //! - Statistics updates
 
 pub mod inflict_table;
+pub mod resolve;
+pub mod stats;
 
-use crate::items::{Item, OwnsItems};
+#[cfg(test)]
+mod tests;
+
+// Re-exports for public API.
+pub use resolve::{AttackContestOutcome, attack_contest};
+pub use stats::update_stats;
+
+// Re-export for tests (pub(crate) items from resolve).
+pub(crate) use resolve::apply_combat_results;
+
 use crate::messages::{CombatEngagement, CombatOutcome, MessagePayload, TaggedEvent, TributeRef};
 use crate::output::GameOutput;
 use crate::tributes::Tribute;
 use crate::tributes::actions::{AttackOutcome, AttackResult};
-use crate::tributes::combat::inflict_table::{
-    HitSeverity, WeaponKind, lookup_break_mid_swing_inflict, lookup_inflicts,
-};
-use rand::RngExt;
 use rand::prelude::*;
 use shared::combat_beat::{CombatBeat, StressReport, SwingOutcome};
-use shared::messages::ItemRef;
-use std::cmp::Ordering;
 
-/// Build a `TributeRef` from a tribute.
-fn tref(t: &Tribute) -> TributeRef {
-    TributeRef {
-        identifier: t.identifier.clone(),
-        name: t.name.clone(),
-    }
-}
+use resolve::{iref, tref};
 
-/// Build an `ItemRef` from an item.
-fn iref(i: &Item) -> ItemRef {
-    ItemRef {
-        identifier: i.identifier.clone(),
-        name: i.name.clone(),
-    }
-}
-
-/// Build a baseline `CombatBeat` (no wear, no stress) for the current swing.
-/// Callers fill in `outcome` and (eventually) `wear` / `stress` per branch.
-fn new_beat(
-    attacker: &Tribute,
-    target: &Tribute,
-    outcome: SwingOutcome,
-    tuning: &crate::tributes::combat_tuning::CombatTuning,
-) -> CombatBeat {
-    CombatBeat {
-        attacker: tref(attacker),
-        target: tref(target),
-        weapon: attacker
-            .items
-            .iter()
-            .rfind(|i| i.is_weapon() && i.current_durability > 0)
-            .map(iref),
-        shield: target
-            .items
-            .iter()
-            .rfind(|i| i.is_defensive() && i.current_durability > 0)
-            .map(iref),
-        wear: Vec::new(),
-        outcome,
-        stress: StressReport::default(),
-        attacker_stamina_cost: tuning.stamina_cost_attacker,
-        target_stamina_cost: tuning.stamina_cost_target,
-    }
-}
+// ---------------------------------------------------------------------------
+// Tribute::attacks — the combat orchestrator
+// ---------------------------------------------------------------------------
 
 impl Tribute {
     /// Tribute attacks another tribute
@@ -116,63 +82,7 @@ impl Tribute {
 
         // Is the tribute attempting suicide?
         if is_self_attack {
-            // Snapshot before mutation.
-            let damage = self.attributes.strength;
-            // Attack always succeeds
-            self.takes_physical_damage(damage);
-            // Capture violence-stress events as standalone (self-harm is not an engagement).
-            let mut stress_events: Vec<TaggedEvent> = Vec::new();
-            let stress_damage = self.apply_violence_stress(&mut stress_events, tuning);
-            events.extend(stress_events);
-
-            return if self.attributes.health > 0 {
-                let content = GameOutput::TributeSelfHarm(self.name.as_str()).to_string();
-                events.push(TaggedEvent::new(
-                    content,
-                    MessagePayload::TributeWounded {
-                        victim: tref(self),
-                        attacker: None,
-                        hp_lost: damage,
-                    },
-                ));
-                let mut beat = new_beat(
-                    self,
-                    target,
-                    SwingOutcome::SelfAttackWound { damage },
-                    tuning,
-                );
-                beat.target_stamina_cost = 0;
-                beat.stress.stress_damage = stress_damage;
-                if stress_damage > 0 {
-                    beat.stress.stressed = Some(tref(self));
-                }
-                events.push(TaggedEvent::new(
-                    String::new(),
-                    MessagePayload::CombatSwing(beat),
-                ));
-                AttackOutcome::Wound(self.clone(), target.clone())
-            } else {
-                let content = GameOutput::TributeSuicide(self.name.as_str()).to_string();
-                events.push(TaggedEvent::new(
-                    content,
-                    MessagePayload::TributeKilled {
-                        victim: tref(self),
-                        killer: None,
-                        cause: "suicide".into(),
-                    },
-                ));
-                let mut beat = new_beat(self, target, SwingOutcome::Suicide { damage }, tuning);
-                beat.target_stamina_cost = 0;
-                beat.stress.stress_damage = stress_damage;
-                if stress_damage > 0 {
-                    beat.stress.stressed = Some(tref(self));
-                }
-                events.push(TaggedEvent::new(
-                    String::new(),
-                    MessagePayload::CombatSwing(beat),
-                ));
-                AttackOutcome::Kill(self.clone(), target.clone())
-            };
+            return self.handle_self_attack(target, events, tuning);
         }
 
         let tribute_name = self.name.clone();
@@ -553,1614 +463,5 @@ impl Tribute {
         ));
 
         attack_outcome
-    }
-
-    /// Apply violence stress to tribute based on their combat history.
-    ///
-    /// Returns the amount of mental damage applied (0 if below threshold).
-    /// Callers thread this value into `CombatBeat.stress.stress_damage` so
-    /// the typed swing payload can reproduce the trailing horrified line
-    /// rendered in `CombatEngagement.detail_lines`.
-    pub(crate) fn apply_violence_stress(
-        &mut self,
-        events: &mut Vec<TaggedEvent>,
-        tuning: &crate::tributes::combat_tuning::CombatTuning,
-    ) -> u32 {
-        let stress_damage = calculate_violence_stress(
-            self.statistics.kills,
-            self.statistics.wins,
-            self.attributes.sanity,
-            tuning,
-        );
-
-        if stress_damage > 0 {
-            let content =
-                GameOutput::TributeHorrified(self.name.as_str(), stress_damage).to_string();
-            events.push(TaggedEvent::new(
-                content,
-                MessagePayload::SanityBreak {
-                    tribute: tref(self),
-                },
-            ));
-            self.takes_mental_damage(stress_damage);
-        }
-        stress_damage
-    }
-}
-
-/// Calculate stress from violent encounters
-fn calculate_violence_stress(
-    kills: u32,
-    wins: u32,
-    current_sanity: u32,
-    tuning: &crate::tributes::combat_tuning::CombatTuning,
-) -> u32 {
-    let non_kill_wins = wins.saturating_sub(kills);
-
-    let calculated_stress_f64 = if wins > 0 {
-        // Calculate the stress potential based on kills and non-kill wins
-        let raw_stress_potential = (kills as f64 * tuning.kill_stress_contribution)
-            + (non_kill_wins as f64 * tuning.non_kill_win_stress_contribution);
-
-        // Desensitize: the more total wins (violent encounters), the more this raw potential
-        // is "spread out" or reduced. This gives an average stressfulness per encounter.
-        let desensitized_stress_per_encounter = raw_stress_potential / wins as f64;
-
-        // Scale by the tribute's current sanity percentage and apply a final divisor.
-        // Lower sanity means less new stress from these types of events.
-        desensitized_stress_per_encounter
-            * (current_sanity as f64 / tuning.stress_sanity_normalization)
-            / tuning.stress_final_divisor
-    } else {
-        // No wins (and therefore no kills), apply a base stress.
-        tuning.base_stress_no_engagements
-    };
-
-    let rounded_stress = calculated_stress_f64.round();
-
-    // Only apply stress if it's at least 1
-    if rounded_stress >= 1.0 {
-        rounded_stress as u32
-    } else {
-        0
-    }
-}
-
-/// Generate attack data for each tribute.
-/// Each rolls a d20 to decide a basic attack / defense value.
-/// Strength and any weapon are added to the attack roll.
-/// Defense and any shield are added to the defense roll.
-/// If either roll is more than 1.5x the other, that triggers a "decisive" victory.
-/// Natural 1 on attack = critical fumble (attacker takes damage).
-/// Natural 20 on attack = critical hit (triple damage).
-/// Natural 20 on defense = perfect block (defender counters).
-/// Outcome of a single `attack_contest` invocation.
-///
-/// Carries the high-level `AttackResult` plus enough wear/penalty data for
-/// the caller to assemble a `CombatBeat` without re-snapshotting equipment
-/// state.
-pub struct AttackContestOutcome {
-    pub result: AttackResult,
-    /// Wear records emitted in attack-roll order: weapon (if equipped) then
-    /// shield (if equipped). Items that were `Pristine` are omitted.
-    pub wear: Vec<shared::combat_beat::WearReport>,
-    /// Affliction drafts to apply to the target (loser of the contest).
-    pub inflicts: Vec<crate::tributes::AfflictionDraft>,
-    /// Affliction drafts to apply to the attacker (e.g. BreakMidSwing recoil).
-    pub attacker_inflicts: Vec<crate::tributes::AfflictionDraft>,
-}
-
-pub fn attack_contest(
-    attacker: &mut Tribute,
-    target: &mut Tribute,
-    rng: &mut impl Rng,
-    events: &mut Vec<TaggedEvent>,
-    tuning: &crate::tributes::combat_tuning::CombatTuning,
-) -> AttackContestOutcome {
-    use shared::combat_beat::{WearOutcomeReport, WearReport};
-    // Get attack roll and strength modifier
-    let base_attack_roll: i32 = rng.random_range(1..=20); // Base roll
-    let mut attack_roll = base_attack_roll;
-    attack_roll += attacker.attributes.strength as i32; // Add strength
-    {
-        use crate::tributes::stamina_band::stamina_band;
-        use shared::messages::StaminaBand;
-        let band = stamina_band(attacker.stamina, attacker.max_stamina, tuning);
-        let penalty = match band {
-            StaminaBand::Fresh => 0,
-            StaminaBand::Winded => tuning.winded_roll_penalty,
-            StaminaBand::Exhausted => tuning.exhausted_roll_penalty,
-        };
-        attack_roll += penalty;
-    }
-
-    let mut wear: Vec<WearReport> = Vec::new();
-
-    // Track weapon kind and whether it broke for inflict table lookup.
-    let mut weapon_kind = WeaponKind::Unarmed;
-    let mut weapon_broken = false;
-
-    // If the attacker has a weapon, use it
-    let weapon_outcome = if let Some(weapon) = attacker.equipped_weapon_mut() {
-        attack_roll += weapon.effect; // Add weapon damage
-        let outcome = weapon.wear(1);
-        // Defer clone: only Worn/Broken paths need the snapshot.
-        match outcome {
-            crate::items::WearOutcome::Pristine => None,
-            _ => Some((weapon.clone(), outcome)),
-        }
-    } else {
-        None
-    };
-    if let Some((weapon, outcome)) = weapon_outcome {
-        weapon_kind = classify_weapon(&weapon);
-        let attacker_ref = tref(attacker);
-        let item_ref = shared::messages::ItemRef {
-            identifier: weapon.identifier.clone(),
-            name: weapon.name.clone(),
-        };
-        match outcome {
-            crate::items::WearOutcome::Pristine => {}
-            crate::items::WearOutcome::Worn => {
-                let content = GameOutput::WeaponWear(attacker.name.as_str(), weapon.name.as_str())
-                    .to_string();
-                events.push(TaggedEvent::new(
-                    content,
-                    MessagePayload::ItemUsed {
-                        tribute: attacker_ref.clone(),
-                        item: item_ref.clone(),
-                    },
-                ));
-                wear.push(WearReport {
-                    owner: attacker_ref,
-                    item: item_ref,
-                    outcome: WearOutcomeReport::Worn,
-                    forfeited_effect: None,
-                    mid_action_penalty: None,
-                });
-            }
-            crate::items::WearOutcome::Broken => {
-                weapon_broken = true;
-                let content = GameOutput::WeaponBreak(attacker.name.as_str(), weapon.name.as_str())
-                    .to_string();
-                events.push(TaggedEvent::new(
-                    content,
-                    MessagePayload::ItemUsed {
-                        tribute: attacker_ref.clone(),
-                        item: item_ref.clone(),
-                    },
-                ));
-                if let Err(err) = attacker.remove_item(&weapon) {
-                    eprintln!("Failed to remove weapon: {}", err);
-                }
-                // D1 + D2: forfeit the just-applied effect bonus and apply 1d4 penalty.
-                attack_roll -= weapon.effect;
-                let penalty = rng.random_range(1..=4);
-                attack_roll -= penalty;
-                let narration = GameOutput::WeaponShattersMidSwing(
-                    attacker.name.as_str(),
-                    weapon.name.as_str(),
-                    penalty as u32,
-                )
-                .to_string();
-                events.push(TaggedEvent::new(
-                    narration,
-                    MessagePayload::ItemUsed {
-                        tribute: attacker_ref.clone(),
-                        item: item_ref.clone(),
-                    },
-                ));
-                wear.push(WearReport {
-                    owner: attacker_ref,
-                    item: item_ref,
-                    outcome: WearOutcomeReport::Broken,
-                    forfeited_effect: Some(weapon.effect),
-                    mid_action_penalty: Some(penalty),
-                });
-            }
-        }
-    }
-
-    // Get defense roll and defense modifier
-    let base_defense_roll: i32 = rng.random_range(1..=20); // Base roll
-    let mut defense_roll = base_defense_roll;
-    defense_roll += target.attributes.defense as i32; // Add defense
-    {
-        use crate::tributes::stamina_band::stamina_band;
-        use shared::messages::StaminaBand;
-        let band = stamina_band(target.stamina, target.max_stamina, tuning);
-        let penalty = match band {
-            StaminaBand::Fresh => 0,
-            StaminaBand::Winded => tuning.winded_roll_penalty,
-            StaminaBand::Exhausted => tuning.exhausted_roll_penalty,
-        };
-        defense_roll += penalty;
-    }
-
-    // If the defender has a shield, use it
-    let shield_outcome = if let Some(shield) = target.equipped_shield_mut() {
-        defense_roll += shield.effect; // Add shield defense
-        let outcome = shield.wear(1);
-        match outcome {
-            crate::items::WearOutcome::Pristine => None,
-            _ => Some((shield.clone(), outcome)),
-        }
-    } else {
-        None
-    };
-    if let Some((shield, outcome)) = shield_outcome {
-        let target_ref = tref(target);
-        let item_ref = shared::messages::ItemRef {
-            identifier: shield.identifier.clone(),
-            name: shield.name.clone(),
-        };
-        match outcome {
-            crate::items::WearOutcome::Pristine => {}
-            crate::items::WearOutcome::Worn => {
-                let content =
-                    GameOutput::ShieldWear(target.name.as_str(), shield.name.as_str()).to_string();
-                events.push(TaggedEvent::new(
-                    content,
-                    MessagePayload::ItemUsed {
-                        tribute: target_ref.clone(),
-                        item: item_ref.clone(),
-                    },
-                ));
-                wear.push(WearReport {
-                    owner: target_ref,
-                    item: item_ref,
-                    outcome: WearOutcomeReport::Worn,
-                    forfeited_effect: None,
-                    mid_action_penalty: None,
-                });
-            }
-            crate::items::WearOutcome::Broken => {
-                let content =
-                    GameOutput::ShieldBreak(target.name.as_str(), shield.name.as_str()).to_string();
-                events.push(TaggedEvent::new(
-                    content,
-                    MessagePayload::ItemUsed {
-                        tribute: target_ref.clone(),
-                        item: item_ref.clone(),
-                    },
-                ));
-                if let Err(err) = target.remove_item(&shield) {
-                    eprintln!("Failed to remove shield: {}", err);
-                }
-                // D3 mirror: forfeit shield effect + apply 1d4 defense penalty.
-                defense_roll -= shield.effect;
-                let penalty = rng.random_range(1..=4);
-                defense_roll -= penalty;
-                let narration = GameOutput::ShieldShattersMidBlock(
-                    target.name.as_str(),
-                    shield.name.as_str(),
-                    penalty as u32,
-                )
-                .to_string();
-                events.push(TaggedEvent::new(
-                    narration,
-                    MessagePayload::ItemUsed {
-                        tribute: target_ref.clone(),
-                        item: item_ref.clone(),
-                    },
-                ));
-                wear.push(WearReport {
-                    owner: target_ref,
-                    item: item_ref,
-                    outcome: WearOutcomeReport::Broken,
-                    forfeited_effect: Some(shield.effect),
-                    mid_action_penalty: Some(penalty),
-                });
-            }
-        }
-    }
-
-    // Check for critical outcomes based on natural rolls (before modifiers)
-    let result = match (base_attack_roll, base_defense_roll) {
-        (1, _) => AttackResult::CriticalFumble, // Natural 1 on attack - fumble!
-        (20, _) => AttackResult::CriticalHit,   // Natural 20 on attack - crit!
-        (_, 20) => AttackResult::PerfectBlock,  // Natural 20 on defense - perfect block!
-        _ => {
-            // Normal combat resolution - compare attack vs. defense
-            match attack_roll.cmp(&defense_roll) {
-                Ordering::Less => {
-                    // If the defender wins
-                    let difference =
-                        defense_roll as f64 - (attack_roll as f64 * tuning.decisive_win_multiplier);
-                    if difference > 0.0 {
-                        // Defender wins significantly
-                        AttackResult::DefenderWinsDecisively
-                    } else {
-                        AttackResult::DefenderWins
-                    }
-                }
-                Ordering::Equal => AttackResult::Miss, // If they tie
-                Ordering::Greater => {
-                    // If the attacker wins
-                    let difference =
-                        attack_roll as f64 - (defense_roll as f64 * tuning.decisive_win_multiplier);
-
-                    if difference > 0.0 {
-                        // Attacker wins significantly
-                        AttackResult::AttackerWinsDecisively
-                    } else {
-                        AttackResult::AttackerWins
-                    }
-                }
-            }
-        }
-    };
-
-    // D5: CriticalFumble clears attacker-side break-penalty fields on the
-    // recorded wear report. The fumble path's catastrophic self-damage is the
-    // story; piling the snapped-weapon penalty on top is double jeopardy.
-    if matches!(result, AttackResult::CriticalFumble)
-        && let Some(report) = wear.iter_mut().find(|w| {
-            w.outcome == WearOutcomeReport::Broken
-                && w.owner.identifier == tref(attacker).identifier
-        })
-    {
-        report.forfeited_effect = None;
-        report.mid_action_penalty = None;
-    }
-
-    // Phase 3: affliction inflict table lookup.
-    let hit_severity = result_to_hit_severity(&result);
-    let attacker_id = attacker.identifier.as_str();
-    let (inflicts, attacker_inflicts) =
-        if matches!(result, AttackResult::Miss | AttackResult::CriticalFumble) {
-            (Vec::new(), Vec::new())
-        } else {
-            let target_inflicts = lookup_inflicts(weapon_kind, hit_severity, attacker_id, rng);
-            // BreakMidSwing follow-through: when weapon shatters, attacker suffers
-            // a recoil injury in addition to any target inflicts.
-            let attacker_inflicts = if weapon_broken {
-                lookup_break_mid_swing_inflict(weapon_kind, attacker_id, rng)
-                    .into_iter()
-                    .collect()
-            } else {
-                Vec::new()
-            };
-            (target_inflicts, attacker_inflicts)
-        };
-
-    AttackContestOutcome {
-        result,
-        wear,
-        inflicts,
-        attacker_inflicts,
-    }
-}
-
-/// Map an AttackResult to the inflict table's severity band.
-fn result_to_hit_severity(result: &AttackResult) -> HitSeverity {
-    match result {
-        AttackResult::CriticalHit => HitSeverity::Critical,
-        AttackResult::AttackerWinsDecisively
-        | AttackResult::DefenderWinsDecisively
-        | AttackResult::PerfectBlock => HitSeverity::Heavy,
-        AttackResult::AttackerWins | AttackResult::DefenderWins => HitSeverity::Normal,
-        AttackResult::CriticalFumble | AttackResult::Miss => HitSeverity::Normal,
-    }
-}
-
-/// Classify an Item into a WeaponKind for the inflict table.
-fn classify_weapon(item: &Item) -> WeaponKind {
-    let name_lower = item.name.to_lowercase();
-    if name_lower.contains("bow") || name_lower.contains("arrow") || name_lower.contains("spear") {
-        WeaponKind::Ranged
-    } else if name_lower.contains("knife")
-        || name_lower.contains("sword")
-        || name_lower.contains("blade")
-        || name_lower.contains("dagger")
-        || name_lower.contains("axe")
-    {
-        WeaponKind::Bladed
-    } else if name_lower.contains("club")
-        || name_lower.contains("hammer")
-        || name_lower.contains("mace")
-        || name_lower.contains("bat")
-    {
-        WeaponKind::Blunt
-    } else {
-        WeaponKind::Bladed // Default for unknown weapons
-    }
-}
-
-/// Apply the results of a combat encounter.
-/// Adjust statistics and log the result.
-///
-/// Returns the stress damage applied to the winner via
-/// `apply_violence_stress` so the caller can populate
-/// `CombatBeat.stress.stress_damage` for the swing payload.
-pub(crate) fn apply_combat_results(
-    winner: &mut Tribute,
-    loser: &mut Tribute,
-    damage_to_loser: u32,
-    log_event: GameOutput,
-    events: &mut Vec<TaggedEvent>,
-    tuning: &crate::tributes::combat_tuning::CombatTuning,
-) -> u32 {
-    loser.takes_physical_damage(damage_to_loser);
-    loser.statistics.defeats += 1;
-    winner.statistics.wins += 1;
-    let stress_damage = winner.apply_violence_stress(events, tuning);
-    let content = log_event.to_string();
-    events.push(TaggedEvent::new(
-        content,
-        MessagePayload::TributeWounded {
-            victim: tref(loser),
-            attacker: Some(tref(winner)),
-            hp_lost: damage_to_loser,
-        },
-    ));
-    stress_damage
-}
-
-/// Update statistics for a pair of tributes based on the attack result
-pub fn update_stats(attacker: &mut Tribute, defender: &mut Tribute, result: AttackResult) {
-    match result {
-        AttackResult::AttackerWins | AttackResult::AttackerWinsDecisively => {
-            defender.statistics.defeats += 1;
-            attacker.statistics.wins += 1;
-        }
-        AttackResult::DefenderWins | AttackResult::DefenderWinsDecisively => {
-            attacker.statistics.defeats += 1;
-            defender.statistics.wins += 1;
-        }
-        AttackResult::CriticalHit => {
-            // Critical hit is a special attacker win (triple damage)
-            defender.statistics.defeats += 1;
-            attacker.statistics.wins += 1;
-        }
-        AttackResult::PerfectBlock => {
-            // Perfect block is a special defender win (counter-attack)
-            attacker.statistics.defeats += 1;
-            defender.statistics.wins += 1;
-        }
-        AttackResult::CriticalFumble => {
-            // Critical fumble: attacker hurts themselves, counts as draw
-            attacker.statistics.draws += 1;
-            defender.statistics.draws += 1;
-        }
-        AttackResult::Miss => {
-            attacker.statistics.draws += 1;
-            defender.statistics.draws += 1;
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::tributes::Tribute;
-    use crate::tributes::combat_tuning::CombatTuning;
-    use core::convert::Infallible;
-    use rand::SeedableRng;
-    use rand::TryRng;
-    use rand::rngs::SmallRng;
-    use rstest::*;
-
-    #[fixture]
-    fn small_rng() -> SmallRng {
-        SmallRng::seed_from_u64(0)
-    }
-
-    #[rstest]
-    fn attack_contest_win(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 10;
-        target.attributes.defense = 5;
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::AttackerWins);
-    }
-
-    #[rstest]
-    fn attack_contest_win_decisively(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 15;
-        target.attributes.defense = 0;
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::AttackerWinsDecisively);
-    }
-
-    #[rstest]
-    fn attack_contest_lose(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 15;
-        target.attributes.defense = 20;
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::DefenderWins);
-    }
-
-    #[rstest]
-    fn attack_contest_lose_decisively(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 1;
-        target.attributes.defense = 20;
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::DefenderWinsDecisively);
-    }
-
-    #[rstest]
-    fn attack_contest_draw(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 21; // Magic number to make the final scores even
-        target.attributes.defense = 20;
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::Miss);
-    }
-
-    #[rstest]
-    fn attacks_self(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        attacker.attributes.strength = 25;
-        attacker.attributes.sanity = 50;
-        let sanity = 50;
-        let mut target = attacker.clone();
-
-        let outcome = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-        assert_eq!(outcome, AttackOutcome::Wound(attacker.clone(), target));
-        assert!(attacker.attributes.sanity < sanity);
-    }
-
-    #[rstest]
-    fn attacks_self_suicide(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        attacker.attributes.strength = 100;
-        let mut target = attacker.clone();
-
-        let outcome = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-        assert_eq!(outcome, AttackOutcome::Kill(attacker, target));
-    }
-
-    #[rstest]
-    fn attacks_wound(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let sanity = attacker.attributes.sanity;
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 25;
-        target.attributes.defense = 20;
-
-        let result = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-        assert_eq!(
-            result,
-            AttackOutcome::Wound(attacker.clone(), target.clone())
-        );
-        assert_eq!(attacker.statistics.wins, 1);
-        assert_eq!(target.statistics.defeats, 1);
-        assert!(attacker.attributes.sanity < sanity);
-    }
-
-    #[rstest]
-    fn attacks_kill(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 50;
-        target.attributes.defense = 0;
-        target.attributes.health = 10;
-
-        let result = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-        assert!(matches!(result, AttackOutcome::Kill(_, _)));
-        assert_eq!(target.attributes.health, 0);
-        assert_eq!(attacker.statistics.wins, 1);
-        assert_eq!(target.statistics.defeats, 1);
-    }
-
-    #[rstest]
-    fn attacks_miss(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 21; // Magic number to make them draw
-        target.attributes.defense = 20;
-
-        let result = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-        assert_eq!(result, AttackOutcome::Miss(attacker, target));
-    }
-
-    #[rstest]
-    fn attacks_deducts_stamina_costs(mut small_rng: SmallRng) {
-        let tuning = CombatTuning::default();
-        let mut attacker = Tribute::new("A".to_string(), None, None);
-        attacker.stamina = 100;
-        attacker.max_stamina = 100;
-        let mut target = Tribute::new("B".to_string(), None, None);
-        target.stamina = 100;
-        target.max_stamina = 100;
-        let _ = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &tuning,
-        );
-        assert_eq!(attacker.stamina, 100 - tuning.stamina_cost_attacker);
-        assert_eq!(target.stamina, 100 - tuning.stamina_cost_target);
-    }
-
-    #[rstest]
-    fn attacks_saturates_at_zero_when_below_cost(mut small_rng: SmallRng) {
-        let tuning = CombatTuning::default();
-        let mut attacker = Tribute::new("A".to_string(), None, None);
-        attacker.stamina = 5;
-        attacker.max_stamina = 100;
-        let mut target = Tribute::new("B".to_string(), None, None);
-        target.stamina = 3;
-        target.max_stamina = 100;
-        let _ = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &tuning,
-        );
-        assert_eq!(attacker.stamina, 0);
-        assert_eq!(target.stamina, 0);
-    }
-
-    #[test]
-    fn attacker_winded_takes_attack_roll_penalty() {
-        // Find a seed where the Winded penalty produces a different outcome
-        // from a Fresh attacker with otherwise identical setup.
-        let tuning = CombatTuning::default();
-        for seed in 0..200u64 {
-            let mut a_fresh = Tribute::new("AF".to_string(), None, None);
-            a_fresh.stamina = 100;
-            a_fresh.max_stamina = 100;
-            a_fresh.attributes.strength = 5;
-            let mut a_winded = Tribute::new("AW".to_string(), None, None);
-            a_winded.stamina = 70; // post-cost: 45 → Winded
-            a_winded.max_stamina = 100;
-            a_winded.attributes.strength = 5;
-            let mut t1 = Tribute::new("T1".to_string(), None, None);
-            t1.stamina = 100;
-            t1.max_stamina = 100;
-            t1.attributes.health = 100;
-            t1.attributes.defense = 10;
-            let mut t2 = Tribute::new("T2".to_string(), None, None);
-            t2.stamina = 100;
-            t2.max_stamina = 100;
-            t2.attributes.health = 100;
-            t2.attributes.defense = 10;
-
-            let mut rng_a = SmallRng::seed_from_u64(seed);
-            let mut rng_b = SmallRng::seed_from_u64(seed);
-            let out_a = a_fresh.attacks(
-                &mut t1,
-                &mut rng_a,
-                &mut Vec::new(),
-                shared::messages::Phase::Day,
-                &tuning,
-            );
-            let out_b = a_winded.attacks(
-                &mut t2,
-                &mut rng_b,
-                &mut Vec::new(),
-                shared::messages::Phase::Day,
-                &tuning,
-            );
-
-            if t1.attributes.health != t2.attributes.health {
-                return; // Found a seed where penalty changes outcome
-            }
-            let _ = (out_a, out_b);
-        }
-        panic!("No seed in 0..200 produced different outcomes for fresh vs winded attacker");
-    }
-
-    #[rstest]
-    fn combat_beat_carries_stamina_costs(mut small_rng: SmallRng) {
-        let tuning = CombatTuning::default();
-        let mut attacker = Tribute::new("A".to_string(), None, None);
-        attacker.stamina = 100;
-        attacker.max_stamina = 100;
-        let mut target = Tribute::new("B".to_string(), None, None);
-        target.stamina = 100;
-        target.max_stamina = 100;
-        let mut events = Vec::new();
-        let _ = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut events,
-            shared::messages::Phase::Day,
-            &tuning,
-        );
-        let beats: Vec<_> = events
-            .iter()
-            .filter_map(|e| match &e.payload {
-                MessagePayload::CombatSwing(b) => Some(b),
-                _ => None,
-            })
-            .collect();
-        assert_eq!(beats.len(), 1);
-        assert_eq!(beats[0].attacker_stamina_cost, tuning.stamina_cost_attacker);
-        assert_eq!(beats[0].target_stamina_cost, tuning.stamina_cost_target);
-    }
-
-    #[rstest]
-    fn test_critical_hit() {
-        // Use a custom RNG that always returns the high bits needed for
-        // `random_range(1..=20)` to produce 20 under rand 0.9's algorithm.
-        struct CritRng;
-        impl TryRng for CritRng {
-            type Error = Infallible;
-            fn try_next_u32(&mut self) -> Result<u32, Infallible> {
-                Ok(0xF333_3334)
-            }
-            fn try_next_u64(&mut self) -> Result<u64, Infallible> {
-                Ok((0xF333_3334u64 << 32) | 0xF333_3334u64)
-            }
-            fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), Infallible> {
-                for byte in dest.iter_mut() {
-                    *byte = 0xFF;
-                }
-                Ok(())
-            }
-        }
-
-        let mut crit_rng = CritRng;
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 10;
-        target.attributes.health = 100;
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut crit_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::CriticalHit);
-    }
-
-    #[rstest]
-    fn test_critical_fumble() {
-        // Use a custom RNG that returns 0 so `random_range(1..=20)` yields 1.
-        struct FumbleRng;
-        impl TryRng for FumbleRng {
-            type Error = Infallible;
-            fn try_next_u32(&mut self) -> Result<u32, Infallible> {
-                Ok(0)
-            }
-            fn try_next_u64(&mut self) -> Result<u64, Infallible> {
-                Ok(0)
-            }
-            fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), Infallible> {
-                for byte in dest.iter_mut() {
-                    *byte = 0;
-                }
-                Ok(())
-            }
-        }
-
-        let mut fumble_rng = FumbleRng;
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut fumble_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::CriticalFumble);
-    }
-
-    #[rstest]
-    fn test_perfect_block() {
-        // First call (attacker roll): 0x7333_3334 → random_range(1..=20) == 10
-        // Second call (defender roll): 0xF333_3334 → random_range(1..=20) == 20
-        struct BlockRng {
-            call_count: std::cell::Cell<usize>,
-        }
-        impl BlockRng {
-            fn new() -> Self {
-                BlockRng {
-                    call_count: std::cell::Cell::new(0),
-                }
-            }
-        }
-        impl TryRng for BlockRng {
-            type Error = Infallible;
-            fn try_next_u32(&mut self) -> Result<u32, Infallible> {
-                let count = self.call_count.get();
-                self.call_count.set(count + 1);
-                Ok(if count == 0 { 0x7333_3334 } else { 0xF333_3334 })
-            }
-            fn try_next_u64(&mut self) -> Result<u64, Infallible> {
-                Ok(self.try_next_u32().unwrap() as u64)
-            }
-            fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), Infallible> {
-                for byte in dest.iter_mut() {
-                    *byte = self.try_next_u32().unwrap() as u8;
-                }
-                Ok(())
-            }
-        }
-
-        let mut block_rng = BlockRng::new();
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        let result = attack_contest(
-            &mut attacker,
-            &mut target,
-            &mut block_rng,
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        )
-        .result;
-        assert_eq!(result, AttackResult::PerfectBlock);
-    }
-
-    #[rstest]
-    fn test_critical_hit_triple_damage(_small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 20;
-        target.attributes.health = 100;
-        let initial_health = target.attributes.health;
-        let damage = attacker.attributes.strength * 3;
-
-        // Manually test the damage application for critical hit
-        apply_combat_results(
-            &mut attacker,
-            &mut target,
-            damage, // Triple damage
-            GameOutput::TributeAttackWin("Katniss", "Peeta"),
-            &mut Vec::new(),
-            &CombatTuning::default(),
-        );
-
-        // Verify triple damage was applied
-        assert_eq!(target.attributes.health, initial_health - damage);
-    }
-
-    #[rstest]
-    fn test_fumble_self_damage(_small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        attacker.attributes.health = 100;
-        let initial_health = attacker.attributes.health;
-
-        // Simulate fumble damage
-        attacker.takes_physical_damage(5);
-
-        assert_eq!(attacker.attributes.health, initial_health - 5);
-    }
-
-    /// Regression test for the clone-mutation bug where weapon wear was
-    /// applied to a cloned item from `weapons()` and silently lost.
-    /// A weapon with durability 5 must survive 3 attack contests with
-    /// reduced durability and remain in the attacker's inventory.
-    #[rstest]
-    fn weapon_survives_multiple_combats(mut small_rng: SmallRng) {
-        use crate::items::{Attribute, Item, ItemRarity, ItemType};
-
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        let weapon = Item::new(
-            "Test Bow",
-            ItemType::Weapon,
-            ItemRarity::Rare,
-            5,
-            Attribute::Strength,
-            3,
-        );
-        let weapon_id = weapon.identifier.clone();
-        attacker.items.push(weapon);
-
-        for _ in 0..3 {
-            attack_contest(
-                &mut attacker,
-                &mut target,
-                &mut small_rng,
-                &mut Vec::new(),
-                &CombatTuning::default(),
-            );
-        }
-
-        let stored = attacker
-            .items
-            .iter()
-            .find(|i| i.identifier == weapon_id)
-            .expect("weapon should still be in inventory after 3 combats");
-        assert_eq!(stored.max_durability, 5);
-        assert_eq!(
-            stored.current_durability, 2,
-            "weapon should have been worn 3 times (5 - 3 = 2)"
-        );
-    }
-
-    #[rstest]
-    fn attacks_target_killed_records_killer_id() {
-        // When the attacker kills the target, target.recently_killed_by must
-        // be set to the attacker's id so the cycle can attribute the death.
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 100;
-        target.attributes.health = 1;
-        target.attributes.defense = 0;
-        let attacker_id = attacker.id;
-
-        // Use a deterministic RNG; with strength=100 vs defense=0 the
-        // attacker reliably wins or crit-hits and the 1hp target dies.
-        let mut rng = SmallRng::seed_from_u64(1);
-        let _ = attacker.attacks(
-            &mut target,
-            &mut rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-
-        assert_eq!(
-            target.attributes.health, 0,
-            "target should be dead in this scenario"
-        );
-        assert_eq!(
-            target.status,
-            crate::tributes::statuses::TributeStatus::RecentlyDead
-        );
-        assert_eq!(
-            target.recently_killed_by,
-            Some(attacker_id),
-            "killer id must be recorded on the deceased"
-        );
-        assert!(
-            attacker.recently_killed_by.is_none(),
-            "attacker is alive; their field must remain None"
-        );
-    }
-
-    #[rstest]
-    fn attacks_attacker_killed_records_target_id() {
-        // When the target's counter kills the attacker (e.g. perfect block),
-        // attacker.recently_killed_by must point to the target.
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.health = 1;
-        attacker.attributes.strength = 0;
-        attacker.attributes.defense = 0;
-        target.attributes.strength = 100;
-        target.attributes.defense = 100;
-        let target_id = target.id;
-
-        let mut rng = SmallRng::seed_from_u64(2);
-        let _ = attacker.attacks(
-            &mut target,
-            &mut rng,
-            &mut Vec::new(),
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-
-        if attacker.attributes.health == 0 {
-            assert_eq!(
-                attacker.status,
-                crate::tributes::statuses::TributeStatus::RecentlyDead
-            );
-            assert_eq!(
-                attacker.recently_killed_by,
-                Some(target_id),
-                "killer id must be recorded on the deceased attacker"
-            );
-        } else {
-            // The seed didn't produce a kill; not a failure of attribution
-            // logic, just RNG. Re-skip rather than flake.
-            eprintln!("seed did not produce attacker death; skipping attribution check");
-        }
-    }
-
-    /// Contract test: a real two-tribute engagement emits exactly one
-    /// `MessagePayload::Combat` `TaggedEvent`, with the attacker/target
-    /// names and a recognised outcome.
-    #[rstest]
-    fn attacks_emits_one_combat_taggedevent(mut small_rng: SmallRng) {
-        let mut attacker = Tribute::new("Katniss".to_string(), None, None);
-        let mut target = Tribute::new("Peeta".to_string(), None, None);
-
-        attacker.attributes.strength = 50;
-        target.attributes.defense = 0;
-        target.attributes.health = 10;
-
-        let mut events: Vec<TaggedEvent> = Vec::new();
-        attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut events,
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-
-        let combat_events: Vec<_> = events
-            .iter()
-            .filter(|e| matches!(e.payload, MessagePayload::Combat(_)))
-            .collect();
-        assert_eq!(
-            combat_events.len(),
-            1,
-            "exactly one Combat payload per attacks() call"
-        );
-
-        if let MessagePayload::Combat(eng) = &combat_events[0].payload {
-            assert_eq!(eng.attacker.name, attacker.name);
-            assert_eq!(eng.target.name, target.name);
-            assert!(matches!(
-                eng.outcome,
-                CombatOutcome::Killed
-                    | CombatOutcome::Wounded
-                    | CombatOutcome::TargetFled
-                    | CombatOutcome::AttackerFled
-                    | CombatOutcome::Stalemate
-            ));
-        } else {
-            panic!("expected Combat payload");
-        }
-    }
-
-    #[rstest]
-    fn attacks_emits_one_combat_swing_per_call(small_rng: SmallRng) {
-        // Every call to attacks() produces exactly one MessagePayload::CombatSwing,
-        // regardless of which branch (miss / wound / kill / fumble / self-attack).
-        for seed in 0..16u64 {
-            let mut events: Vec<TaggedEvent> = Vec::new();
-            let mut attacker = Tribute::new(format!("A{seed}"), None, None);
-            let mut target = Tribute::new(format!("T{seed}"), None, None);
-            let mut rng = SmallRng::seed_from_u64(seed);
-            let _ = attacker.attacks(
-                &mut target,
-                &mut rng,
-                &mut events,
-                shared::messages::Phase::Day,
-                &CombatTuning::default(),
-            );
-
-            let swings: Vec<_> = events
-                .iter()
-                .filter(|e| matches!(e.payload, MessagePayload::CombatSwing(_)))
-                .collect();
-            assert_eq!(
-                swings.len(),
-                1,
-                "seed {seed}: expected exactly one CombatSwing payload, got {}",
-                swings.len()
-            );
-
-            if let MessagePayload::CombatSwing(beat) = &swings[0].payload {
-                assert_eq!(beat.attacker.name, attacker.name);
-                assert_eq!(beat.target.name, target.name);
-            }
-        }
-        // Touch the rstest fixture to silence unused warnings.
-        let _ = small_rng;
-    }
-
-    #[rstest]
-    fn self_attack_emits_one_combat_swing(mut small_rng: SmallRng) {
-        let mut tribute = Tribute::new("Solo".to_string(), None, None);
-        let mut clone = tribute.clone();
-        let mut events: Vec<TaggedEvent> = Vec::new();
-        let _ = tribute.attacks(
-            &mut clone,
-            &mut small_rng,
-            &mut events,
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-        let swings: usize = events
-            .iter()
-            .filter(|e| matches!(e.payload, MessagePayload::CombatSwing(_)))
-            .count();
-        assert_eq!(swings, 1, "self-attack must emit exactly one CombatSwing");
-    }
-
-    /// Construct a weapon with the given effect and durability=1 so a single
-    /// `wear(1)` call breaks it.
-    fn brittle_weapon(effect: i32) -> crate::items::Item {
-        crate::items::Item::new(
-            "Glass Sword",
-            crate::items::ItemType::Weapon,
-            crate::items::ItemRarity::Common,
-            1,
-            crate::items::Attribute::Strength,
-            effect,
-        )
-    }
-
-    /// Construct a shield with the given effect and durability=1.
-    fn brittle_shield(effect: i32) -> crate::items::Item {
-        crate::items::Item::new(
-            "Glass Buckler",
-            crate::items::ItemType::Weapon,
-            crate::items::ItemRarity::Common,
-            1,
-            crate::items::Attribute::Defense,
-            effect,
-        )
-    }
-
-    #[test]
-    fn weapon_break_records_forfeit_and_penalty_on_beat() {
-        let mut attacker = Tribute::new("Atk".into(), None, None);
-        attacker.attributes.strength = 10;
-        let weapon = brittle_weapon(5);
-        attacker.add_item(weapon.clone());
-
-        let mut target = Tribute::new("Tgt".into(), None, None);
-        target.attributes.defense = 5;
-
-        let mut events: Vec<TaggedEvent> = Vec::new();
-        let mut rng = SmallRng::seed_from_u64(42);
-        let _ = attacker.attacks(
-            &mut target,
-            &mut rng,
-            &mut events,
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-
-        let beat = events
-            .iter()
-            .find_map(|e| match &e.payload {
-                MessagePayload::CombatSwing(b) => Some(b),
-                _ => None,
-            })
-            .expect("expected one CombatSwing emission");
-        let weapon_wear = beat
-            .wear
-            .iter()
-            .find(|w| w.owner.identifier == beat.attacker.identifier)
-            .expect("expected a wear report for the attacker's weapon");
-
-        assert_eq!(
-            weapon_wear.outcome,
-            shared::combat_beat::WearOutcomeReport::Broken
-        );
-        assert_eq!(weapon_wear.forfeited_effect, Some(5));
-        let penalty = weapon_wear.mid_action_penalty.expect("penalty must fire");
-        assert!(
-            (1..=4).contains(&penalty),
-            "penalty must be 1..=4, got {}",
-            penalty
-        );
-    }
-
-    #[test]
-    fn shield_break_records_forfeit_and_penalty_on_beat() {
-        let mut attacker = Tribute::new("Atk".into(), None, None);
-        attacker.attributes.strength = 10;
-
-        let mut target = Tribute::new("Tgt".into(), None, None);
-        target.attributes.defense = 5;
-        let shield = brittle_shield(4);
-        target.add_item(shield.clone());
-
-        let mut events: Vec<TaggedEvent> = Vec::new();
-        let mut rng = SmallRng::seed_from_u64(7);
-        let _ = attacker.attacks(
-            &mut target,
-            &mut rng,
-            &mut events,
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-
-        let beat = events
-            .iter()
-            .find_map(|e| match &e.payload {
-                MessagePayload::CombatSwing(b) => Some(b),
-                _ => None,
-            })
-            .expect("expected one CombatSwing emission");
-        let shield_wear = beat
-            .wear
-            .iter()
-            .find(|w| w.owner.identifier == beat.target.identifier)
-            .expect("expected a wear report for the target's shield");
-
-        assert_eq!(
-            shield_wear.outcome,
-            shared::combat_beat::WearOutcomeReport::Broken
-        );
-        assert_eq!(shield_wear.forfeited_effect, Some(4));
-        let penalty = shield_wear.mid_action_penalty.expect("penalty must fire");
-        assert!((1..=4).contains(&penalty), "penalty was {}", penalty);
-    }
-
-    #[test]
-    fn fumble_clears_attacker_break_penalty_on_beat() {
-        // Hunt for a seed where the attacker's swing both fumbles AND breaks
-        // the brittle weapon. Per design D5 the attacker-side break-penalty
-        // fields must be cleared on fumble for clean narration.
-        for seed in 0u64..2_000 {
-            let mut attacker = Tribute::new("Atk".into(), None, None);
-            attacker.attributes.strength = 10;
-            let weapon = brittle_weapon(5);
-            attacker.add_item(weapon.clone());
-
-            let mut target = Tribute::new("Tgt".into(), None, None);
-            target.attributes.defense = 5;
-
-            let mut events: Vec<TaggedEvent> = Vec::new();
-            let mut rng = SmallRng::seed_from_u64(seed);
-            let _ = attacker.attacks(
-                &mut target,
-                &mut rng,
-                &mut events,
-                shared::messages::Phase::Day,
-                &CombatTuning::default(),
-            );
-
-            let beat = match events.iter().find_map(|e| match &e.payload {
-                MessagePayload::CombatSwing(b) => Some(b),
-                _ => None,
-            }) {
-                Some(b) => b,
-                None => continue,
-            };
-
-            let is_fumble = matches!(
-                beat.outcome,
-                shared::combat_beat::SwingOutcome::FumbleSurvive { .. }
-                    | shared::combat_beat::SwingOutcome::FumbleDeath { .. }
-            );
-            let weapon_wear = beat
-                .wear
-                .iter()
-                .find(|w| w.owner.identifier == beat.attacker.identifier);
-
-            if is_fumble
-                && weapon_wear
-                    .map(|w| w.outcome == shared::combat_beat::WearOutcomeReport::Broken)
-                    .unwrap_or(false)
-            {
-                let w = weapon_wear.unwrap();
-                assert_eq!(
-                    w.forfeited_effect, None,
-                    "D5: fumble must clear forfeited_effect"
-                );
-                assert_eq!(
-                    w.mid_action_penalty, None,
-                    "D5: fumble must clear mid_action_penalty"
-                );
-                return;
-            }
-        }
-        panic!("no seed in 0..2000 produced a fumble + weapon break combo; widen the search");
-    }
-
-    #[test]
-    fn unarmed_unshielded_emits_no_break_penalty() {
-        let mut attacker = Tribute::new("Atk".into(), None, None);
-        attacker.attributes.strength = 10;
-        let mut target = Tribute::new("Tgt".into(), None, None);
-        target.attributes.defense = 5;
-
-        let mut events: Vec<TaggedEvent> = Vec::new();
-        let mut rng = SmallRng::seed_from_u64(123);
-        let _ = attacker.attacks(
-            &mut target,
-            &mut rng,
-            &mut events,
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-
-        let beat = events
-            .iter()
-            .find_map(|e| match &e.payload {
-                MessagePayload::CombatSwing(b) => Some(b),
-                _ => None,
-            })
-            .expect("expected one CombatSwing emission");
-        for w in &beat.wear {
-            assert_eq!(w.forfeited_effect, None);
-            assert_eq!(w.mid_action_penalty, None);
-        }
-    }
-
-    /// Parity: wear lines rendered from `CombatBeat.wear` via
-    /// `CombatBeatExt::to_log_lines` must match the wear lines that
-    /// `attack_contest` flattens into `CombatEngagement.detail_lines`.
-    /// This locks `CombatBeat.wear` as the single source of truth for wear
-    /// narration once consumers migrate off `detail_lines`.
-    #[test]
-    fn combat_beat_wear_matches_engagement_detail_wear_lines() {
-        use crate::tributes::combat_beat::CombatBeatExt;
-        // Both attacker weapon and target shield are brittle so we exercise
-        // the full wear (Worn or Broken) emission paths in one swing.
-        for seed in 0..32u64 {
-            let mut attacker = Tribute::new(format!("Atk{seed}"), None, None);
-            attacker.attributes.strength = 10;
-            attacker.add_item(brittle_weapon(5));
-            let mut target = Tribute::new(format!("Tgt{seed}"), None, None);
-            target.attributes.defense = 5;
-            target.add_item(brittle_shield(4));
-
-            let mut events: Vec<TaggedEvent> = Vec::new();
-            let mut rng = SmallRng::seed_from_u64(seed);
-            let _ = attacker.attacks(
-                &mut target,
-                &mut rng,
-                &mut events,
-                shared::messages::Phase::Day,
-                &CombatTuning::default(),
-            );
-
-            // Find the engagement (skip seeds that didn't produce one, e.g.
-            // pure fumble paths emit a standalone TributeWounded instead and
-            // never write detail_lines for the legacy timeline to consume).
-            let Some(detail_lines) = events.iter().find_map(|e| match &e.payload {
-                MessagePayload::Combat(eng) => Some(eng.detail_lines.clone()),
-                _ => None,
-            }) else {
-                continue;
-            };
-            let beat = events
-                .iter()
-                .find_map(|e| match &e.payload {
-                    MessagePayload::CombatSwing(b) => Some(b),
-                    _ => None,
-                })
-                .expect("expected one CombatSwing per attacks() call");
-
-            // Wear-related substrings we want to track.
-            let is_wear_line = |s: &str| {
-                s.contains("starting to wear out")
-                    || s.contains("breaks")
-                    || s.contains("shatters mid-swing")
-                    || s.contains("shatters mid-block")
-            };
-
-            let beat_wear_lines: Vec<String> = beat
-                .to_log_lines()
-                .into_iter()
-                .filter(|s| is_wear_line(s))
-                .collect();
-            let detail_wear_lines: Vec<String> = detail_lines
-                .into_iter()
-                .filter(|s| is_wear_line(s))
-                .collect();
-
-            assert_eq!(
-                beat_wear_lines, detail_wear_lines,
-                "seed {seed}: wear lines from CombatBeat must match detail_lines"
-            );
-        }
-    }
-
-    /// A swing that wins (and therefore triggers `apply_violence_stress`) on
-    /// a tribute already credited with prior wins must record the resulting
-    /// stress on the swing's `CombatBeat.stress.stress_damage`.
-    #[test]
-    fn combat_swing_records_stress_damage() {
-        // Arrange: attacker has prior wins+kills so violence-stress is non-zero.
-        let mut attacker = Tribute::new("Atk".into(), None, None);
-        attacker.attributes.strength = 50;
-        attacker.statistics.wins = 5;
-        attacker.statistics.kills = 5;
-
-        let mut target = Tribute::new("Tgt".into(), None, None);
-        target.attributes.defense = 0;
-        target.attributes.health = 1;
-
-        let mut events: Vec<TaggedEvent> = Vec::new();
-        let mut rng = SmallRng::seed_from_u64(1);
-        let _ = attacker.attacks(
-            &mut target,
-            &mut rng,
-            &mut events,
-            shared::messages::Phase::Day,
-            &CombatTuning::default(),
-        );
-
-        let beat = events
-            .iter()
-            .find_map(|e| match &e.payload {
-                MessagePayload::CombatSwing(b) => Some(b),
-                _ => None,
-            })
-            .expect("expected one CombatSwing emission");
-        assert!(
-            beat.stress.stress_damage > 0,
-            "expected non-zero stress on the beat; got {}",
-            beat.stress.stress_damage
-        );
-    }
-
-    /// Parity: the optional trailing horrified line that
-    /// `CombatBeatExt::to_log_lines` renders from `stress.stress_damage`
-    /// must match the horrified line that `apply_violence_stress` flattened
-    /// into `CombatEngagement.detail_lines`.
-    #[test]
-    fn combat_beat_stress_matches_engagement_horrified_line() {
-        use crate::tributes::combat_beat::CombatBeatExt;
-        for seed in 0..32u64 {
-            let mut attacker = Tribute::new(format!("Atk{seed}"), None, None);
-            attacker.attributes.strength = 50;
-            attacker.statistics.wins = 5;
-            attacker.statistics.kills = 5;
-
-            let mut target = Tribute::new(format!("Tgt{seed}"), None, None);
-            target.attributes.defense = 0;
-            target.attributes.health = 100;
-
-            let mut events: Vec<TaggedEvent> = Vec::new();
-            let mut rng = SmallRng::seed_from_u64(seed);
-            let _ = attacker.attacks(
-                &mut target,
-                &mut rng,
-                &mut events,
-                shared::messages::Phase::Day,
-                &CombatTuning::default(),
-            );
-
-            let Some(detail_lines) = events.iter().find_map(|e| match &e.payload {
-                MessagePayload::Combat(eng) => Some(eng.detail_lines.clone()),
-                _ => None,
-            }) else {
-                continue;
-            };
-            let beat = events
-                .iter()
-                .find_map(|e| match &e.payload {
-                    MessagePayload::CombatSwing(b) => Some(b),
-                    _ => None,
-                })
-                .expect("expected one CombatSwing per attacks() call");
-
-            let beat_horrified: Vec<String> = beat
-                .to_log_lines()
-                .into_iter()
-                .filter(|s| s.contains("horrified"))
-                .collect();
-            let detail_horrified: Vec<String> = detail_lines
-                .into_iter()
-                .filter(|s| s.contains("horrified"))
-                .collect();
-
-            assert_eq!(
-                beat_horrified, detail_horrified,
-                "seed {seed}: horrified lines from CombatBeat must match detail_lines"
-            );
-        }
-    }
-
-    #[rstest]
-    fn attacking_sleeping_target_emits_ambush_wake(mut small_rng: SmallRng) {
-        // Spec §6.4 PR2c.2 (bd-1zju): hitting a sleeping tribute must wake
-        // them with `WakeReason::Interrupted { Ambush }` BEFORE damage
-        // resolution. The ambush still lands; we just verify the wake
-        // event precedes any TributeWounded / Killed event in `events`.
-        let mut attacker = Tribute::new("Cato".to_string(), None, None);
-        attacker.attributes.strength = 10;
-        let mut target = Tribute::new("Rue".to_string(), None, None);
-        target.sleeping = true;
-        target.sleep_remaining = 4;
-        target.cycles_awake = 12;
-
-        let mut events: Vec<TaggedEvent> = Vec::new();
-        let _ = attacker.attacks(
-            &mut target,
-            &mut small_rng,
-            &mut events,
-            shared::messages::Phase::Night,
-            &CombatTuning::default(),
-        );
-
-        assert!(!target.sleeping, "ambushed target must wake");
-        assert_eq!(target.sleep_remaining, 0);
-        assert_eq!(target.cycles_awake, 0);
-
-        let woke_idx = events
-            .iter()
-            .position(|e| {
-                matches!(
-                    &e.payload,
-                    MessagePayload::TributeWoke {
-                        reason: shared::messages::WakeReason::Interrupted {
-                            event: shared::messages::InterruptionKind::Ambush { .. },
-                        },
-                        ..
-                    }
-                )
-            })
-            .expect("expected one ambush TributeWoke event");
-        assert_eq!(
-            woke_idx, 0,
-            "ambush wake must be emitted before damage resolution"
-        );
     }
 }
